+++ src/components/data-management/MetadataColumn.jsx
@@ -2,16 +2,51 @@ import React from 'react';
 import { Space, Input } from 'antd';
 import { useDispatch } from 'react-redux';
 import PropTypes from 'prop-types';
-import { rules } from '../../utils/validateInputs';
+import validateInputs, { rules } from '../../utils/validateInputs';
 import EditableField from '../EditableField';
 import MetadataEditor from './MetadataEditor';
 import { DEFAULT_NA } from '../../redux/reducers/projects/initialState';
+import { metadataNameToKey } from '../../utils/data-management/metadataUtils';
 
 import {
   updateMetadataTrack,
 } from '../../redux/actions/projects';
 
-const MetadataColumn = (props) => {
+const MetadataColumnTitle = (props) => {
+  const {
+    name, sampleNames, activeProjectUuid, deleteMetadataColumn, setCells,
+  } = props;
+  const key = metadataNameToKey(name);
+
+  const validationParams = {
+    existingNames: sampleNames,
+  };
+
+  return (
+    <MetadataTitle
+      name={name}
+      validateInput={
+        (newName, metadataNameValidation) => validateInputs(
+          newName, metadataNameValidation, validationParams,
+        ).isValid
+      }
+      setCells={setCells}
+      deleteMetadataColumn={deleteMetadataColumn}
+      key={key}
+      activeProjectUuid={activeProjectUuid}
+    />
+  );
+};
+
+MetadataColumnTitle.propTypes = {
+  name: PropTypes.string.isRequired,
+  sampleNames: PropTypes.instanceOf(Set).isRequired,
+  setCells: PropTypes.func.isRequired,
+  deleteMetadataColumn: PropTypes.func.isRequired,
+  activeProjectUuid: PropTypes.string.isRequired,
+};
+
+const MetadataTitle = (props) => {
   const dispatch = useDispatch();
   const {
     name, validateInput, setCells, deleteMetadataColumn, key, activeProjectUuid,
@@ -46,7 +81,7 @@ const MetadataColumn = (props) => {
     </Space>
   );
 };
-MetadataColumn.propTypes = {
+MetadataTitle.propTypes = {
   name: PropTypes.string.isRequired,
   validateInput: PropTypes.func.isRequired,
   setCells: PropTypes.func.isRequired,
@@ -54,4 +89,4 @@ MetadataColumn.propTypes = {
   key: PropTypes.string.isRequired,
   activeProjectUuid: PropTypes.string.isRequired,
 };
-export default MetadataColumn;
+export default MetadataColumnTitle;
